+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Updated
+
+- [BMP-507](https://makeitapp.atlassian.net/browse/BMP-507): updated glogger logs format
+
 ## 1.0.0 - 10/12/2019
 
 - Initial Release 🎉🎉🎉